+++ src/features/editor/components/fixed-menu/index.tsx
@@ -4,7 +4,7 @@ import { Button } from "@/components/ui/button";
 import { showToast } from "@/lib/toast";
 import { cn } from "@/lib/utils";
 import { INITIAL_CONTENT } from "../../lib/initial-content";
-import { markdownSerializer } from "../../lib/to_markdown";
+import { markdownSerializer } from "../../lib/to-markdown";
 import MarkdownPasteDialog from "../markdown-paste-dialog";
 
 type Props = {
